refactor(dpp): extract shared try_from_schema parsing helpers
The document-type parser generations (v0/v1/v2) carried near-identical
copies of name validation, schema depth/size accounting, meta-schema
selection + compilation, keeps*History flag extraction, the core
property/index parsing body, and the doctype-level aggregate keyword
handling. Extract them into try_from_schema/common as parameter-pure
helpers: nothing in common reads the platform-version tables; each
generation passes its own constants (v1 supplies the one genuinely
version-varying input, the document_type_schema >= 2 keeps*History
gate, since upstream v2 delegates core parsing to v1).

Zero behavior change: the full dpp suite passes with an identical
test-name set (3806 passed before and after), no test was added or
removed, and the only test-module edits pin each generation's tests
to that generation's own protocol versions (v0->PV8, v1->PV11 with
the meta-schema axis at PV12, v2->PV13) instead of latest(), so the
guards keep targeting the generation they were written for as new
protocol versions land.

Net -1115 lines across the parser generations.
