+++ packages/rs-dpp/src/data_contract/document_type/class_methods/try_from_schema/mod.rs
@@ -15,6 +15,7 @@ use platform_value::{Identifier, Value};
 use platform_version::version::PlatformVersion;
 use std::collections::{BTreeMap, BTreeSet};
 
+mod common;
 mod v0;
 mod v1;
 mod v2;
